requirements.txt
pyramid
waitress
sqlalchemy
zope.sqlalchemy
pyramid_tm
pyramid_jinja2
